backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
python-dotenv==1.0.1
django-debug-toolbar==4.2.0
django-debug-toolbar-force==0.2
psycopg2-binary==2.9.9
drf-yasg==1.21.7
setuptools==69.0.3
Pillow==10.2.0
django-cors-headers==4.3.1
celery[redis]==5.3.6
flower==2.0.1
celery-singleton==0.3.1
djoser==2.2.2